Refactor: Move checkMultiSigPubKey to TransactionValidation
