+++ core/src/test/java/bisq/core/trade/validation/DepositTxValidationTest.java
@@ -34,7 +34,6 @@ import bisq.network.p2p.NodeAddress;
 import bisq.common.crypto.CryptoException;
 import bisq.common.crypto.PubKeyRing;
 import bisq.common.crypto.Sig;
-import bisq.common.util.Utilities;
 
 import com.google.protobuf.ByteString;
 
@@ -82,71 +81,6 @@ public class DepositTxValidationTest {
 
 
 
-    /* --------------------------------------------------------------------- */
-    // DepositTx checkMultiSigPubKey
-    /* --------------------------------------------------------------------- */
-
-    @Test
-    void checkMultiSigPubKeyAcceptsCompressedPublicKey() {
-        byte[] multiSigPubKey = new ECKey().getPubKey();
-
-        assertEquals(33, multiSigPubKey.length);
-        assertSame(multiSigPubKey, DepositTxValidation.checkMultiSigPubKey(multiSigPubKey));
-    }
-
-    @Test
-    void checkMultiSigPubKeyRejectsNullPublicKey() {
-        assertThrows(NullPointerException.class, () -> DepositTxValidation.checkMultiSigPubKey(null));
-    }
-
-    @Test
-    void checkMultiSigPubKeyRejectsUncompressedPublicKey() {
-        byte[] uncompressedPubKey = new ECKey().decompress().getPubKey();
-
-        assertEquals(65, uncompressedPubKey.length);
-        assertThrows(IllegalArgumentException.class, () -> DepositTxValidation.checkMultiSigPubKey(uncompressedPubKey));
-    }
-
-    @Test
-    void checkMultiSigPubKeyRejectsMalformedCompressedPublicKey() {
-        byte[] malformedCompressedPubKey = new byte[33];
-        Arrays.fill(malformedCompressedPubKey, (byte) 0xff);
-        malformedCompressedPubKey[0] = 0x02;
-
-        assertThrows(IllegalArgumentException.class,
-                () -> DepositTxValidation.checkMultiSigPubKey(malformedCompressedPubKey));
-    }
-
-    @Test
-    void checkMultiSigPubKeyAcceptsValidCompressedCurvePoints() {
-        // These deterministic encodings exercise x-coordinates where x^3 + 7 is a quadratic residue mod the
-        // secp256k1 field prime, so both compressed y-parity prefixes map to valid curve points.
-        String[] validEncodings = {
-                "020000000000000000000000000000000000000000000000000000000000000001",
-                "020000000000000000000000000000000000000000000000000000000000000002",
-                "020000000000000000000000000000000000000000000000000000000000000003",
-                "020000000000000000000000000000000000000000000000000000000000000004",
-                "020000000000000000000000000000000000000000000000000000000000000006",
-                "020000000000000000000000000000000000000000000000000000000000000008",
-                "02000000000000000000000000000000000000000000000000000000000000000c",
-                "02000000000000000000000000000000000000000000000000000000000000000d",
-                "02000000000000000000000000000000000000000000000000000000000000000e",
-                "030000000000000000000000000000000000000000000000000000000000000001",
-                "030000000000000000000000000000000000000000000000000000000000000002",
-                "030000000000000000000000000000000000000000000000000000000000000003",
-                "030000000000000000000000000000000000000000000000000000000000000004",
-                "030000000000000000000000000000000000000000000000000000000000000006",
-                "030000000000000000000000000000000000000000000000000000000000000008",
-                "03000000000000000000000000000000000000000000000000000000000000000c",
-                "03000000000000000000000000000000000000000000000000000000000000000d",
-                "03000000000000000000000000000000000000000000000000000000000000000e"
-        };
-
-        for (String validEncoding : validEncodings) {
-            byte[] multiSigPubKey = Utilities.decodeFromHex(validEncoding);
-            assertDoesNotThrow(() -> DepositTxValidation.checkMultiSigPubKey(multiSigPubKey), validEncoding);
-        }
-    }
 
     /* --------------------------------------------------------------------- */
     // InputsForDepositTxRequest
@@ -209,41 +143,6 @@ public class DepositTxValidationTest {
     }
 
 
-
-
-    /* --------------------------------------------------------------------- */
-    //
-    /* --------------------------------------------------------------------- */
-
-
-    @Test
-    void checkMultiSigPubKeyRejectsInvalidCompressedCurvePoints() {
-        // These x-coordinates do not produce a quadratic residue for x^3 + 7 mod the secp256k1 field prime,
-        // so neither compressed y-parity prefix can decompress to a valid curve point.
-        String[] invalidEncodings = {
-                "020000000000000000000000000000000000000000000000000000000000000000",
-                "020000000000000000000000000000000000000000000000000000000000000005",
-                "020000000000000000000000000000000000000000000000000000000000000007",
-                "020000000000000000000000000000000000000000000000000000000000000009",
-                "02000000000000000000000000000000000000000000000000000000000000000a",
-                "02000000000000000000000000000000000000000000000000000000000000000b",
-                "02000000000000000000000000000000000000000000000000000000000000000f",
-                "030000000000000000000000000000000000000000000000000000000000000000",
-                "030000000000000000000000000000000000000000000000000000000000000005",
-                "030000000000000000000000000000000000000000000000000000000000000007",
-                "030000000000000000000000000000000000000000000000000000000000000009",
-                "03000000000000000000000000000000000000000000000000000000000000000a",
-                "03000000000000000000000000000000000000000000000000000000000000000b",
-                "03000000000000000000000000000000000000000000000000000000000000000f"
-        };
-
-        for (String invalidEncoding : invalidEncodings) {
-            byte[] multiSigPubKey = Utilities.decodeFromHex(invalidEncoding);
-            assertThrows(IllegalArgumentException.class,
-                    () -> DepositTxValidation.checkMultiSigPubKey(multiSigPubKey),
-                    invalidEncoding);
-        }
-    }
 
 
     @Test
