+++ core/src/test/java/bisq/core/trade/validation/TransactionValidationTest.java
@@ -17,6 +17,8 @@
 
 package bisq.core.trade.validation;
 
+import bisq.common.util.Utilities;
+
 import org.bitcoinj.core.ECKey;
 import org.bitcoinj.core.SegwitAddress;
 import org.bitcoinj.core.Transaction;
@@ -31,10 +33,7 @@ import java.util.Locale;
 import org.junit.jupiter.api.Test;
 
 import static bisq.core.trade.validation.TradeValidationTestUtils.btcWalletService;
-import static org.junit.jupiter.api.Assertions.assertArrayEquals;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertSame;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class TransactionValidationTest {
     @Test
@@ -226,5 +225,102 @@ public class TransactionValidationTest {
 
         assertThrows(IllegalArgumentException.class, () -> TransactionValidation.checkDerEncodedEcdsaSignature(bitcoinSignature));
     }
+
+
+    /* --------------------------------------------------------------------- */
+    // DepositTx checkMultiSigPubKey
+    /* --------------------------------------------------------------------- */
+
+    @Test
+    void checkMultiSigPubKeyAcceptsCompressedPublicKey() {
+        byte[] multiSigPubKey = new ECKey().getPubKey();
+
+        assertEquals(33, multiSigPubKey.length);
+        assertSame(multiSigPubKey, TransactionValidation.checkMultiSigPubKey(multiSigPubKey));
+    }
+
+    @Test
+    void checkMultiSigPubKeyRejectsNullPublicKey() {
+        assertThrows(NullPointerException.class, () -> TransactionValidation.checkMultiSigPubKey(null));
+    }
+
+    @Test
+    void checkMultiSigPubKeyRejectsUncompressedPublicKey() {
+        byte[] uncompressedPubKey = new ECKey().decompress().getPubKey();
+
+        assertEquals(65, uncompressedPubKey.length);
+        assertThrows(IllegalArgumentException.class, () -> TransactionValidation.checkMultiSigPubKey(uncompressedPubKey));
+    }
+
+    @Test
+    void checkMultiSigPubKeyRejectsMalformedCompressedPublicKey() {
+        byte[] malformedCompressedPubKey = new byte[33];
+        Arrays.fill(malformedCompressedPubKey, (byte) 0xff);
+        malformedCompressedPubKey[0] = 0x02;
+
+        assertThrows(IllegalArgumentException.class,
+                () -> TransactionValidation.checkMultiSigPubKey(malformedCompressedPubKey));
+    }
+
+    @Test
+    void checkMultiSigPubKeyAcceptsValidCompressedCurvePoints() {
+        // These deterministic encodings exercise x-coordinates where x^3 + 7 is a quadratic residue mod the
+        // secp256k1 field prime, so both compressed y-parity prefixes map to valid curve points.
+        String[] validEncodings = {
+                "020000000000000000000000000000000000000000000000000000000000000001",
+                "020000000000000000000000000000000000000000000000000000000000000002",
+                "020000000000000000000000000000000000000000000000000000000000000003",
+                "020000000000000000000000000000000000000000000000000000000000000004",
+                "020000000000000000000000000000000000000000000000000000000000000006",
+                "020000000000000000000000000000000000000000000000000000000000000008",
+                "02000000000000000000000000000000000000000000000000000000000000000c",
+                "02000000000000000000000000000000000000000000000000000000000000000d",
+                "02000000000000000000000000000000000000000000000000000000000000000e",
+                "030000000000000000000000000000000000000000000000000000000000000001",
+                "030000000000000000000000000000000000000000000000000000000000000002",
+                "030000000000000000000000000000000000000000000000000000000000000003",
+                "030000000000000000000000000000000000000000000000000000000000000004",
+                "030000000000000000000000000000000000000000000000000000000000000006",
+                "030000000000000000000000000000000000000000000000000000000000000008",
+                "03000000000000000000000000000000000000000000000000000000000000000c",
+                "03000000000000000000000000000000000000000000000000000000000000000d",
+                "03000000000000000000000000000000000000000000000000000000000000000e"
+        };
+
+        for (String validEncoding : validEncodings) {
+            byte[] multiSigPubKey = Utilities.decodeFromHex(validEncoding);
+            assertDoesNotThrow(() -> TransactionValidation.checkMultiSigPubKey(multiSigPubKey), validEncoding);
+        }
+    }
+
+    @Test
+    void checkMultiSigPubKeyRejectsInvalidCompressedCurvePoints() {
+        // These x-coordinates do not produce a quadratic residue for x^3 + 7 mod the secp256k1 field prime,
+        // so neither compressed y-parity prefix can decompress to a valid curve point.
+        String[] invalidEncodings = {
+                "020000000000000000000000000000000000000000000000000000000000000000",
+                "020000000000000000000000000000000000000000000000000000000000000005",
+                "020000000000000000000000000000000000000000000000000000000000000007",
+                "020000000000000000000000000000000000000000000000000000000000000009",
+                "02000000000000000000000000000000000000000000000000000000000000000a",
+                "02000000000000000000000000000000000000000000000000000000000000000b",
+                "02000000000000000000000000000000000000000000000000000000000000000f",
+                "030000000000000000000000000000000000000000000000000000000000000000",
+                "030000000000000000000000000000000000000000000000000000000000000005",
+                "030000000000000000000000000000000000000000000000000000000000000007",
+                "030000000000000000000000000000000000000000000000000000000000000009",
+                "03000000000000000000000000000000000000000000000000000000000000000a",
+                "03000000000000000000000000000000000000000000000000000000000000000b",
+                "03000000000000000000000000000000000000000000000000000000000000000f"
+        };
+
+        for (String invalidEncoding : invalidEncodings) {
+            byte[] multiSigPubKey = Utilities.decodeFromHex(invalidEncoding);
+            assertThrows(IllegalArgumentException.class,
+                    () -> TransactionValidation.checkMultiSigPubKey(multiSigPubKey),
+                    invalidEncoding);
+        }
+    }
+
 
 }
